refactor(calendar): UniversalCalendar sous 300 lignes — éclatement complet (H8)
Éclate UniversalCalendar.vue (1154 l) en orchestrateur (228 l) + :
- composable useCalendarView.js (état vue FullCalendar, options, navigation, sync)
- CalendarNavigation.vue / CalendarViewSelector.vue / CalendarView.vue / CalendarLegend.vue
- partials _calendar-fc-dark.scss + _calendar-shell-dark.scss (dark mode global verbatim)

CSS déplacé verbatim (clair scoped dans chaque sous-composant, dark non-scoped en
partials @use). Parité : mêmes appels services (via useCalendarEvents), mêmes
classes/textes, getApi() relayé par CalendarView. Tests: 26 verts. vite build OK.

Dettes pré-existantes conservées à l'identique et documentées :
- variable router (useRouter) déclarée sans usage dans l'orchestrateur
- règle media .filters-content (CSS scoped) ne cible plus le DOM enfant depuis #28bis
