+++ tests/unit/UniversalCalendar.test.js
@@ -1,12 +1,13 @@
 /**
- * Test de montage (smoke) de UniversalCalendar.vue (#28bis / G8 — Agent B).
- * Le composant utilise useRouter(), le composable useCalendarEvents (appels réseau),
- * FullCalendar et les sous-composants CalendarFilters / EventDetailModal.
+ * Test de montage (smoke) de UniversalCalendar.vue (#28bis / G8 / H8).
+ * Depuis H8, l'orchestrateur compose CalendarNavigation, CalendarFilters,
+ * CalendarView (qui rend FullCalendar), CalendarLegend et EventDetailModal, et
+ * délègue l'état de la vue à useCalendarView.
  * On mocke :
  *  - vue-router (useRouter)
  *  - @/composables/useCalendarEvents → refs neutres + loadEvents/refreshData no-op (zéro réseau)
- * et on stub FullCalendar, CalendarFilters, EventDetailModal, ContentLoader.
- * Assertion : montage sans erreur (racine .calendar-container présente).
+ * et on stub les sous-composants présentationnels.
+ * Assertion : montage sans erreur (racine .calendar-container présente) + loadEvents au montage.
  */
 import { mount } from '@vue/test-utils'
 import { ref } from 'vue'
@@ -39,10 +40,11 @@ function mountCalendar(props = {}) {
     props: { userRole: 'teacher', userId: 1, ...props },
     global: {
       stubs: {
-        FullCalendar: true,
+        CalendarNavigation: true,
         CalendarFilters: true,
-        EventDetailModal: true,
-        ContentLoader: true
+        CalendarView: true,
+        CalendarLegend: true,
+        EventDetailModal: true
       }
     }
   })
